pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "crunch-node"
version = "0.1.86"
description = "Runtime engine for Crunch coordinator nodes"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "alembic>=1.13,<2",
    "fastapi>=0.100.0,<1",
    "model-runner-client>=0.10.0",
    "psycopg2-binary>=2.9,<3",

    "python-binance>=1.0,<2",
    "requests>=2.28,<3",
    "sqlmodel>=0.0.14,<1",
    "uvicorn>=0.20,<1",
]

[project.optional-dependencies]
mongodb = ["pymongo>=4.0,<5"]
parquet = ["pyarrow>=15.0,<19"]

[project.scripts]
crunch-node = "crunch_node.cli:main"

[project.urls]
Homepage = "https://github.com/crunchdao/coordinator-node-starter"
Repository = "https://github.com/crunchdao/coordinator-node-starter"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
exclude = [
    "**/.venv/**",
    "**/__pycache__/**",
    "**/.pytest_cache/**",
    "**/*.pyc",
    "**/.DS_Store",
    "**/.ruff_cache/**",
]

[tool.hatch.build.targets.sdist]
# Exclude VCS-tracked scaffold/packs, then force-include them cleanly
# to avoid duplicates from .gitignore exceptions (e.g. !.env.example).
exclude = ["scaffold/**", "packs/**"]
force-include = { "scaffold" = "scaffold", "packs" = "packs" }

[tool.hatch.build.targets.wheel]
packages = ["crunch_node"]

[tool.hatch.build.targets.wheel.force-include]
"scaffold" = "crunch_node/_templates/scaffold"
"packs" = "crunch_node/_templates/packs"

[tool.uv]
package = true

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = [
    "E402",   # module-level imports (some files do sys.path setup first)
    "E501",   # line length handled by formatter
    "F401",   # re-exports in __init__.py are intentional
    "F811",   # redefinitions in tests (helper class overrides)
    "F821",   # forward-ref type annotations resolved at runtime
    "F841",   # unused variables (often intentional _ pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["crunch_node"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0",
    "ruff>=0.9",
]
